acculynx-agent/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# AccuLynx API client
httpx>=0.27.0

# AI (Claude)
anthropic>=0.40.0

# Messaging
twilio>=9.0.0
sendgrid>=6.11.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dateutil>=2.9.0
rich>=13.0.0
pyyaml>=6.0
pyjwt>=2.8.0
jinja2>=3.1.0
python-multipart>=0.0.9
greenlet>=3.0.0
